requirements.txt
Babel==2.3.4
certifi==2016.9.26
click==6.6
cssselect==1.0.0
Flask==0.11.1
Flask-Babel==0.11.1
flask-mongoengine==0.7.5
Flask-WTF==0.12
gevent==1.1.2
greenlet==0.4.10
itsdangerous==0.24
Jinja2==2.8
lorem-ipsum-generator==0.3
lxml==3.6.4
MarkupSafe==0.23
mongoengine==0.10.6
Pillow==3.4.2
pymongo==3.3.1
pytz==2016.7
PyYAML==3.12
requests==2.11.1
six==1.10.0
transliterate==1.8.1
ujson==1.35
uWSGI==2.0.15
Werkzeug==0.11.11
WTForms==2.1
